requirements.txt
# Seismic-BD project dependencies
# Python 3.13+ required
# Install with: pip install -r requirements.txt

# Core data science
pandas>=2.2
numpy>=2.0
scipy>=1.14
statsmodels>=0.14

# Visualization
matplotlib>=3.9
seaborn>=0.13

# Geospatial
geopandas>=1.0
shapely>=2.0
contextily>=1.6

# Document parsing
python-docx>=1.1
docx2txt>=0.8

# Performance / serialization
pyarrow>=17.0

# Date/string utilities
python-dateutil>=2.9
regex>=2024.0

# Notebooks
jupyter>=1.1
ipykernel>=6.29
